Add SOC data
Add option to use unnormalized raw datasets from a file.
Also add an option to normalize the raw data according to how many reads are in a single gene.

Options to center around the mean or use raw data instead of normalized now take effect on the underlying data matrix.  This means that any MDS results calculated from that matrix are also affected.  This is a good thing as it allow s you to try out various scenarios and see the effect on the P value calculated from the modified matrix.