Modify morrislab.R to issue a warning if no matching alignments are found, not stop!
Modify profiling/server.R to include the dataset name along with the
description when presenting a list to the user.

When looking up alignments in the database from server.R, if no
alignments are found under the Refseq name, try using the the name as
specified by the user.  This will allow a mixture of alignment to be
found, some of which are under the common name and some of which are
undet the RefSeq name.