Code to make a union gene model
This code combines exons from genes that have the same common name.
This is flawed for a number of reasons.  It seems to be overly
conservative in some ways, and not conservative enough in others.
Although it avoids combining "isoforms" that are located on different
strands or chromosomes, it will combine widely separated genes if they
have the same name, Also, depending on the curation source it may not
combine genes that really should be combined, because they don't have
an identical common name.

A "better" approach might be to combine overlapping regions?  That has
its own problems.